specify task num at cmd line for interactive job

+++ scripts/submit_array.sh
@@ -1,22 +1,27 @@
 #!/bin/bash
-# usage: bash submit_array.sh -g hg19 -r /SAN/colcc/alex_work/samples_analysis -c "" -s configuration/2021-06_icgc_cancer_vs_normal.txt
+# usage: bash submit_array.sh -g hg19 -r /SAN/colcc/alex_work/samples_analysis -c _nochr -s configuration/2021-06_icgc_cancer_vs_normal.txt
+#
+# or specify a SGE_TASK_ID, in which case run interactively rather than submitting, in which case
+# usage: bash submit_array.sh -g hg19 -r /SAN/colcc/alex_work/samples_analysis -c _nochr -s configuration/2021-06_icgc_cancer_vs_normal.txt -t 57 >stdout.txt 2>stderr.txt &
 
 # some defaults
 GENOME="hg19" # default; must be hg19 or grch38
 CHR_SUFFIX="_nochr" # default; must be either "_nochr" OR "_chr"; genome files (fasta, bed) are of the form grch38_nochr*, hg19_chr*, etc
 RESULTS_BASE_DIR=/SAN/colcc/alex_work/samples_analysis
 SAMPLE_METADATA=configuration/2021-06_icgc_cancer_vs_normal.txt
+SGE_TASK_ID=-1
 
 # set variables from arguments
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )" # https://stackoverflow.com/questions/59895/how-can-i-get-the-source-directory-of-a-bash-script-from-within-the-script-itsel
 PPLN_BASE_DIR=$SCRIPT_DIR/..
-while getopts g:r:s: flag
+while getopts g:r:s:t: flag
 do
     case "${flag}" in
         g) GENOME=${OPTARG};;            # 'hg19' or 'grch38'
-        c) CHR_SUFFIX=${OPTARG};;      # '_nochr' or '' if chromosomes are specified without or with 'chr' respectively; used to choose appropriately named genomic bed and fasta files
+        c) CHR_SUFFIX=${OPTARG};;        # '_chr' or '_nochr' if chromosomes are specified with or without 'chr' respectively; used to choose appropriately named genomic bed and fasta files
         r) RESULTS_BASE_DIR=${OPTARG};;  # base directory for analysis results
         s) SAMPLE_METADATA=${OPTARG};;   # sample metadata file
+        t) SGE_TASK_ID=${OPTARG};;       # run a specific task (line number in the file, after removal of successfully completed jobs)
     esac
 done
 echo "Genome: $GENOME";
@@ -29,25 +34,34 @@ then
     mkdir $RESULTS_BASE_DIR
 fi
 
-# check for finished jobs, with lohhla outputs
-# identify pt ids (subdirs within RESULTS_BASE_DIR) with completed figures
-# use found pt ids to search first col in configuration/xx.txt; create file of grep patterns; empty file ok; do not repeat analysis in case of success
+# if a job is specified (only to be done if running interactively), analyse chosen sample and exit the script
+if [ $SGE_TASK_ID -gt -1 ]
+then
+    . ./job.sh # extra . means run-in-this-env
+    exit 0
+fi
+
+# otherwise, check for finished jobs, with lohhla outputs
 SAMPLE_METADATA_TMP=$SAMPLE_METADATA".tmp"
 SUCCESS_IDS_TMP=$SAMPLE_METADATA".success"
-ls $RESULTS_BASE_DIR/*/lohhla/results/Figures/*.HLA.pdf | \
-    perl -ne 'chomp; s/.lohhla.results.Figures..*$//; s/^.*\///; print "^$_\t\n"' > $SUCCESS_IDS_TMP
-grep -Ev -f $SUCCESS_IDS_TMP $SAMPLE_METADATA > $SAMPLE_METADATA_TMP
-rm $SUCCESS_IDS_TMP
-# note, upon lohhla failure, Figures dirs will not exist, so no deletion necessary before rerunning job.
+if [ `ls $RESULTS_BASE_DIR/*/lohhla/results/Figures/*.HLA.pdf | grep -c ^` -gt 0 ]
+then
+    # identify pt ids (subdirs within RESULTS_BASE_DIR) with completed figures
+    # use found pt ids to search first col in configuration/xx.txt; create file of grep patterns; empty file ok; do not repeat analysis in case of success
+    ls $RESULTS_BASE_DIR/*/lohhla/results/Figures/*.HLA.pdf | \
+       perl -ne 'chomp; s/.lohhla.results.Figures..*$//; s/^.*\///; print "^$_\t\n"' > $SUCCESS_IDS_TMP
+    grep -Ev -f $SUCCESS_IDS_TMP $SAMPLE_METADATA > $SAMPLE_METADATA_TMP
+    rm $SUCCESS_IDS_TMP
+else
+    cp $SAMPLE_METADATA $SAMPLE_METADATA_TMP
+fi
 
 # calculate number of tasks required
 SAMPLES_LINES=($(cat $SAMPLE_METADATA_TMP | cut -f 1)) # includes a header
 NUM_LINES=${#SAMPLES_LINES[@]} # array length
 TASK_COUNT=$((NUM_LINES-1))
 
-echo "GENOME=$GENOME SAMPLE_METADATA=$SAMPLE_METADATA_TMP TASK_COUNT=$TASK_COUNT "
-#SGE_TASK_ID=2 && . ./job.sh # extra . means run-in-this-env
-#exit 0
+echo "GENOME=$GENOME SAMPLE_METADATA=$SAMPLE_METADATA_TMP RESULTS_BASE_DIR=$RESULTS_BASE_DIR TASK_COUNT=$TASK_COUNT SGE_TASK_ID=$SGE_TASK_ID "
 
 # qsub the array job
 # Request 1 core, 5 hour runtime (esp if 'chr' must be cleared from bam files), 20GB RAM, job array, 2 tasks run recurrently
